avoid double copy of data
Replaces EncodePublishMessage and EncodeDataMessage, which copy the
data into the message buffer, with EncodePublishMessagePrefix and
EncodeDataMessagePrefix. Then the server and SDK use writeev with the
prefix and data to avoid an extra copy.

Running the benchmarks shows a 35% reduction in the publish scenario and
a 65% redunction in the subscribe scenario.